refactor(user): cherry-pick 후 :test-support 기반으로 테스트 인프라 재작성
- ServiceTest/IntegrationTestSupport: 로컬 TestContainerSupport 제거, test-support 상속
- IntegrationTest 어노테이션: classes=UserModuleTestApplication.class 명시
- ServiceTestConfig: 외부 모듈 포트(ProfanityChecker, RoomInvitationValidator 등) mock 추가
- application-test.yml: application-test-base.yml import 방식으로 전환
- 이동된 테스트 파일: coffeeshout.support.app.* import → 모듈 자체 기반 클래스로 교체
- UserRegistrationServiceTest: ProfanityChecker mock 행동 설정 추가
