+++ backend/user/src/test/java/coffeeshout/ServiceTest.java
@@ -1,7 +1,8 @@
 package coffeeshout;
 
 import coffeeshout.config.ServiceTestConfig;
-import coffeeshout.fixture.TestContainerSupport;
+import coffeeshout.global.nickname.ProfanityChecker;
+import coffeeshout.support.TestContainerSupport;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.context.annotation.Import;
@@ -9,7 +10,7 @@ import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.transaction.annotation.Transactional;
 
-@SpringBootTest
+@SpringBootTest(classes = UserModuleTestApplication.class)
 @Import(ServiceTestConfig.class)
 @ActiveProfiles("test")
 @Transactional
@@ -17,4 +18,7 @@ public abstract class ServiceTest extends TestContainerSupport {
 
     @MockitoBean
     protected ApplicationEventPublisher eventPublisher;
+
+    @MockitoBean
+    protected ProfanityChecker profanityChecker;
 }
