+++ backend/user/src/test/java/coffeeshout/fixture/IntegrationTestSupport.java
@@ -1,5 +1,6 @@
 package coffeeshout.fixture;
 
+import coffeeshout.support.TestContainerSupport;
 import coffeeshout.support.test.IntegrationTest;
 import org.junit.jupiter.api.BeforeEach;
