+++ backend/user/src/test/resources/application-test.yml
@@ -1,66 +1,3 @@
 spring:
-  datasource:
-    url: jdbc:h2:mem:testdb;MODE=MySQL;DATABASE_TO_LOWER=TRUE;DEFAULT_NULL_ORDERING=HIGH
-    hikari:
-      maximum-pool-size: 5
-      minimum-idle: 0
-      idle-timeout: 10000
-  security:
-    oauth2:
-      client:
-        registration:
-          google:
-            client-id: test-google-client-id
-            client-secret: test-google-client-secret
-            scope: profile,email
-          kakao:
-            client-id: test-kakao-client-id
-            client-secret: test-kakao-client-secret
-            authorization-grant-type: authorization_code
-            redirect-uri: "{baseUrl}/login/oauth2/code/{registrationId}"
-            client-authentication-method: client_secret_post
-            scope: profile_nickname,account_email
-          naver:
-            client-id: test-naver-client-id
-            client-secret: test-naver-client-secret
-            authorization-grant-type: authorization_code
-            redirect-uri: "{baseUrl}/login/oauth2/code/{registrationId}"
-            scope: name,email
-  jpa:
-    hibernate:
-      ddl-auto: create
-    properties:
-      hibernate:
-        format_sql: true
-        show_sql: false
-  lifecycle:
-    timeout-per-shutdown-phase: 1s
-  flyway:
-    enabled: false
-  data:
-    redis:
-      ssl:
-        enabled: false
-  autoconfigure:
-    exclude:
-      - org.redisson.spring.starter.RedissonAutoConfigurationV2
-
-server:
-  shutdown: immediate
-
-redisson:
-  enabled: true
-
-logging:
-  level:
-    coffeeshout: DEBUG
-  pattern:
-    console: "%d{yyyy-MM-dd HH:mm:ss} [%thread] %-5level %logger{36} - %msg%n"
-
-user:
-  jwt:
-    secret: "test-secret-key-must-be-at-least-256-bits-long-for-hs256-algorithm"
-    access-token-expiration-seconds: 1800
-    refresh-token-expiration-seconds: 1209600
-  oauth:
-    frontend-redirect-uri: "http://localhost:3000/auth/callback"
+  config:
+    import: "classpath:application-test-base.yml"
